// client/src/app/components/drawing-surface/drawing-surface-attributes.ts
export class DrawingSurfaceAttributes {
    toolImplemented: boolean;
    fillTool: boolean;
    showGrid: boolean;
    pipetteTool: boolean;
    gridSquareSize: number;
    gridTransparency: number;
    onSurface: boolean;
    drawingCreated: boolean;
    color: string;
    width: string;
    height: string;

}